+++ CHANGELOG.md
@@ -4,6 +4,8 @@
   where there previously were none, but all handlers should have already been
   returning a `Response` or `Future<Response>`.
 * Remove `HijackCallback` and `OnHijackCallback` typedefs.
+* **Breaking**: Change type of `onHijack` in the `Request` constructor to take
+  an argument of `StreamChannel`.
 
 ## 0.6.8